pyproject.toml
[project]
name = "tiny-mcf"
version = "0.1.0"
description = "Tiny moral-foundations vignettes eval (negation + self/other) for steering checkpoints."
requires-python = ">=3.11"
dependencies = [
    "transformers>=4.45",
    "torch",
    "accelerate",
    "pandas",
    "loguru",
    "python-dotenv",
    "tyro",
    "openai>=1.40",
    "httpx",
    "tqdm",
    "tabulate",
    "openrouter-wrapper",
    "datasets",
]

[tool.uv.sources]
openrouter-wrapper = { path = "../daily-dilemmas-self/docs/openrouter_wrapper", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tinymcf"]
